Bug fix
Noticed that when text is modified in the autocomplete input text to
add a letter or extend the selected word, when focus is lost on the
input text, the changed text reverts back to the original entry.  This
checks to see if the updated value is an extension to the previously
selected word, and if so, updates it.